database/sku/requirements.txt
# Зависимости для проекта database/sku
# База данных спецификаций товаров Wookiee

# База данных
psycopg2-binary>=2.9.9      # PostgreSQL драйвер
sqlalchemy>=2.0.0           # ORM

# Работа с Excel
pandas>=2.0.0               # Обработка данных
openpyxl>=3.1.0             # Чтение Excel файлов

# Миграции (опционально)
alembic>=1.13.0             # Миграции БД

# Утилиты
python-dotenv>=1.0.0        # Переменные окружения из .env

# Bitrix24 интеграция
requests>=2.31.0            # HTTP клиент
